* Updated project files for Visual C++ 2008. * Cleaned up some warnings.
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%401072

+++ src/lasheader.cpp
@@ -404,7 +404,7 @@ LASHeader::PointFormat LASHeader::GetDataFormatId() const
 
 void LASHeader::SetDataFormatId(LASHeader::PointFormat v)
 {
-    m_dataFormatId = v;
+    m_dataFormatId = static_cast<uint8_t>(v);
 
     if (ePointFormat0 == m_dataFormatId)
         m_dataRecordLen = ePointSize0;
